% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofsimsData-package.R
\docType{data}
\name{testImage}
\alias{testImage}
\title{Example ToF-SIMS data}
\format{A MassImage object}
\usage{
data(tofsimsData)
}
\value{
MassImage object
}
\description{
A dataset containing a MassImage recorded on a
Ulvac-Phi TRIFT-II ToF-SIMS. The .RAW data file was
imported using
\code{tofsimsImage<-MassImage('ulvacrawpeaks','filename', PeakList=tofsimsSpectra)}.
The sample is a freeze-dried transversal poplar wood
section of 100 micrometer thickness.
}
\keyword{dataset}

