% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofsimsData-package.R
\docType{data}
\name{testSpectra}
\alias{testSpectra}
\title{Example ToF-SIMS data}
\format{A MassSpectra object}
\usage{
data(tofsimsData)
}
\value{
MassSpectra object
}
\description{
A dataset containing a MassSpectra recorded on a
Ulvac-Phi TRIFT-II ToF-SIMS. The .RAW data file was
imported using
\code{tofsimsSpectra<-MassSpectra('ulvacraw','filename')}.
The sample is a freeze-dried transversal poplar wood
section of 100 micrometer thickness.
}
\keyword{dataset}

