% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofsimsData-package.R
\docType{package}
\name{tofsimsData-package}
\alias{tofsimsData-package}
\title{tofsimsData}
\description{
ToF-SIMS Toolbox
}
\details{
\tabular{ll}{
Package: \tab tofsimsData\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 23-10-2014\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes \cr
}

Toolbox for Time-of-Flight Secondary Ion Mass-Spectrometry
(ToF-SIMS) data processing and analysis. The package facilitates
importing of raw data files, loading preprocessed data and a range of
multivariate analysis methods that are most commonly applied in imaging
(ToF-SIMS) mass spectrometry.
}
\author{
Lorenz Gerber \email{lorenz.gerber@slu.se}
}
\keyword{package}

